# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Nsangusa API
  version: 1.0.0
  description: Static language-neutral contract for implemented Spring MVC routes.
    Spring Security supplies the explicitly marked login and logout operations.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://api.example.invalid
  description: Placeholder; configure per environment
...
security:
- sessionCookie: []
- basicAuth: []
paths:
...
  "/api/v1/newsletter/subscriptions":
    post:
      tags:
      - Newsletter
      operationId: requestNewsletterSubscription
      summary: Request double opt-in and send a real confirmation email
      description: >
        Normalizes the email before lookup. An existing subscription is rejected with 400; this legacy signup
        operation is not the enumeration-safe preference-link request. The response never contains the address or
        verification token. SMTP acceptance of the confirmation message does not confirm newsletter consent.
      security: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Phase4NewsletterSubscriptionRequest'
      responses:
        '202':
          description: Confirmation message accepted by the configured mail transport; consent remains pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phase4NewsletterSubscriptionRequested'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/newsletter/confirm":
    post:
      tags:
      - Newsletter
      operationId: confirmNewsletterSubscription
      summary: Explicitly confirm a pending newsletter subscription
      description: >
        Replaces the former GET operation. GET is no longer supported and never changes consent. Requires the
        emailed verification token and browser CSRF protection. Repeating the same valid confirmation after
        confirmation is idempotent. Old verification links cannot reactivate an unsubscribed or suppressed
        subscription.
      security: []
      parameters:
      - $ref: '#/components/parameters/SubscriptionId'
      - $ref: '#/components/parameters/SubscriptionToken'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Consent confirmed, or the same confirmation was already completed.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
  "/api/v1/newsletter/unsubscribe":
    post:
      tags:
      - Newsletter
      operationId: unsubscribeNewsletter
      summary: Withdraw newsletter consent using the dedicated unsubscribe token
      description: >
        Public token-authorized POST used by both the explicit browser control and List-Unsubscribe one-click
        requests. This endpoint is narrowly exempt from CSRF; other newsletter browser mutations are not. GET is
        unsupported. Repeated valid requests are idempotent. Suppression remains in force. The unsubscribe token
        cannot authorize frequency changes.
      security: []
      parameters:
      - $ref: '#/components/parameters/SubscriptionId'
      - $ref: '#/components/parameters/SubscriptionToken'
      requestBody:
        required: false
        description: Optional RFC 8058 one-click form body; authorization remains in the query token.
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                List-Unsubscribe:
                  type: string
                  const: One-Click
      responses:
        '204':
          description: Newsletter consent withdrawn without removing an existing suppression.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  "/api/v1/newsletter/provider-webhooks/{provider}":
    post:
      tags:
      - Newsletter
      operationId: processNewsletterProviderWebhook
      summary: Verify and deduplicate a configured email-provider callback
      description: >
        Verifies the raw body before parsing. Resend requires valid Svix headers with a timestamp within five
        minutes; the explicitly configured local legacy provider uses X-Webhook-Signature HMAC. The path provider
        must match configuration. Repeated external event IDs are ignored. Bounce and complaint callbacks preserve
        suppression. Only a signed delivery callback matched by provider message ID can establish
        deliveryConfirmedAt; SMTP acceptance alone never does. Delivery callbacks without a message identifier are
        rejected. Provider-specific SMTP/callback correlation still requires production qualification.
      security: []
      parameters:
      - $ref: '#/components/parameters/Provider'
      - name: svix-id
        in: header
        description: Required when the configured provider is resend.
        schema:
          type: string
      - name: svix-timestamp
        in: header
        description: Required for resend; Unix seconds within the five-minute tolerance.
        schema:
          type: string
          pattern: ^[0-9]+$
      - name: svix-signature
        in: header
        description: Required for resend; one or more versioned Svix signatures.
        schema:
          type: string
      - name: X-Webhook-Signature
        in: header
        description: Required for the explicitly configured local legacy webhook format.
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: {}
          text/plain:
            schema:
              type: string
      responses:
        '204':
          description: Verified callback processed, or the same external event was already recorded.
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Redis-backed opaque Spring Session cookie
    basicAuth:
      type: http
      scheme: basic
  parameters:
...
    Provider:
      name: provider
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    CsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: Spring Security CSRF token matching the XSRF-TOKEN cookie
      schema:
        type: string
        minLength: 1
...
    SubscriptionId:
      name: id
      in: query
      required: true
      schema:
        "$ref": "#/components/schemas/Identifier"
    SubscriptionToken:
      name: token
      in: query
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Invalid path, query, header, or request-body input
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    Forbidden:
      description: Authentication succeeded but authorization or CSRF validation failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Conflict:
      description: The requested transition conflicts with current resource state
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    TooManyRequests:
      description: Per-client request rate exceeded
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Retry delay in seconds
    ServiceUnavailable:
      description: Rate limiting could not reach its Redis dependency
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    UnexpectedError:
      description: Unexpected API failure
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
    Identifier:
      type: string
      format: uuid
...
    Problem:
      type: object
      additionalProperties: true
      required:
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    Phase4NewsletterFrequency:
      type: string
      enum:
      - immediate
      - daily
      - weekly
      - all
      description: All selects immediate messages and both digest frequencies.
...
    Phase4NewsletterSubscriptionRequest:
      type: object
      required:
      - email
      - consentSource
      - frequency
      properties:
        email:
          type: string
          format: email
          minLength: 1
        consentSource:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9][A-Za-z0-9._:/-]{0,199}$
        frequency:
          $ref: '#/components/schemas/Phase4NewsletterFrequency'
    Phase4NewsletterSubscriptionRequested:
      type: object
      additionalProperties: false
      required:
      - subscriptionId
      - status
      properties:
        subscriptionId:
          $ref: '#/components/schemas/Identifier'
        status:
          type: string
          const: verification_sent
          description: Confirmation message submitted; subscriber consent is not yet confirmed.
...